Rework how large objects are segmented and properly obey SLO min size
This commit unifies the logic handling DLO and SLO segmentation, it also
adds more test cases repeatedly writing and seeking large objects.
The minimum chunk size is now exposed and can also be used with DLOs if
a large number of potentially small objects is not desired (with the
performance penalty of having to retrieve the old segment when writing
over it).

The minimum size for SLOs is also retrived from a /info call and will
override chunkSize and minChunkSize if it's larger.

The implementation of the ReadFrom method was dropped in the writer. The
reason for this is that the old implementation was incorrect sometimes
due to it using the number of written bytes to find out if it had
already finished. If the last segment had the exact same size as the
chunk size it would try writing one more segment, and this segment would
be empty. One way to solve this would be adding a DELETE call if a zero
sized segment was written, but this is ugly and wasteful. Another option
would be to buffer the read data in memory to find out if we're done,
but this is exactly what io.Copy() already do for us, so I could simply
drop the ReadFrom implementation.