Rework tests to avoid shared state between test cases
The main advantage of having independent tests is being able to run a
single test with `go test -run <regex>`. This makes debugging easier and
is a good practice in general.

To acomplish this all test cases were changed to rollback changes on the
swift server. Any individual test case executed should leave no
containers/objects in the server, even in case of failure.