catch "304 Not Modified" response to ObjectOpen() correctly
This is a followup to 05ee6c556644e1f98f247b4b0abc0240bff30428,
in which I didn't consider that a 304 response from Swift will not even
produce a non-nil *ObjectOpenFile. It will simply exit with an error.

This patch allows response status 304 to be detected via
`err == swift.NotModified.`