Handle OVH 429 Too Many Requests for cloud archive
Before being able to retrive an object from OVH Cloud Archive, an object
must be unfreezed. First call and subsequent calls until the object is
available, OVH will return a 429 status and Retry-After header set to a
time in seconds after which the object should be avaible.

This change makes status 429 a testable error, and returns headers, so the
caller can extract the Retry-After header.