Ignore lint errors for strings.Title in test code
strings.Title is broken for certain types of unicode string, however
in this test code we are only ever using on single word ASCII strings
where it works fine. We don't want to take another dependency to fix
(x/text) which is the recommended fix, so we will just ignore the lint
errors since they are false positives here.